R: Extract class to parser
